+++ src/math_MonteCarloSequential.cpp
@@ -24,7 +24,7 @@
 // Prototypes:
 ////////////////////////////////////////////////////////////////////////////////
 
-double runSequentialMonteCarloPICalculation(int n);
+double runSequentialMonteCarloPICalculation(long n);
 
 ////////////////////////////////////////////////////////////////////////////////
 // Globals:
@@ -34,7 +34,7 @@ double runSequentialMonteCarloPICalculation(int n);
 // Procedures:
 ////////////////////////////////////////////////////////////////////////////////
 
-double runSequentialMonteCarloPICalculation(int n) {
+double runSequentialMonteCarloPICalculation(long n) {
   int count = 0;
   std::default_random_engine generator;
   std::uniform_real_distribution<double> distribution(0.0, 1.0);
